Update Buffer Size

+++ source/res/ResFunc.c
@@ -6,7 +6,7 @@
 #include <windows.h>
 
 // Unsure for now - Check the functionality later.
-wchar_t GetResourceStringW(HINSTANCE hInst, int strID, size_t BuffSize){
+wchar_t GetResourceStringW(HINSTANCE hInst, int strID, int BuffSize){
 	static wchar_t Buffer[BuffSize];
 	
 	LoadStringW(hInst, strID, Buffer, sizeof(Buffer) / sizeof(wchar_t));
@@ -14,7 +14,7 @@ wchar_t GetResourceStringW(HINSTANCE hInst, int strID, size_t BuffSize){
 }
 
 // Same as with Wide-char one - Check the functionality later.
-char GetResourceStringA(HINSTANCE hInst, int strID, size_t BuffSize){
+char GetResourceStringA(HINSTANCE hInst, int strID, int BuffSize){
 	static char Buffer[BuffSize];
 	
 	LoadStringA(hInst, strID, Buffer, sizeof(Buffer) / sizeof(Buffer[0]));
